pyproject.toml
[tool.poetry]
name = "tfg-model"
version = "0.1.0"
description = ""
authors = ["Carlos Fernandez <carlos.ferlo@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
implica = "^1.2.2"
jax = "^0.9.0.1"
jraphx = "^0.0.4"
hydra-core = "^1.3.2"
wandb = "^0.25.0"
optax = "0.2.6"
flax = "^0.12.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
